% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLUvalid.R
\docType{package}
\name{CLUvalid}
\alias{CLUvalid}
\alias{CLUvalid-package}
\title{CLUvalid: a package with tools for processing SSP results and external data to enable comparison}
\description{
The CLUvalid package provides three categories of functions and scripts:
preprocessing, visual comparison and numeric comparison.
}
