% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_reclass_matrix.R
\name{create_rcl_LC}
\alias{create_rcl_LC}
\title{Create reclass matrix}
\usage{
create_rcl_LC(t1, t2)
}
\arguments{
\item{t1}{first threshold}

\item{t2}{second threshold}
}
\value{
a matrix with three columns to be used as a reclass table
}
\description{
Create reclass matrix
}
