% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/List_CLU_folders.R
\name{list_CLU_folders}
\alias{list_CLU_folders}
\title{List the locations of the CLU folders}
\usage{
list_CLU_folders(inpath, scenario)
}
\arguments{
\item{x}{the path to the folder containing the scenario folder}

\item{y}{the scenario name}
}
\value{
The locations of the folders as a list
}
\description{
List the locations of the CLU folders
}
\examples{
CLU_folders <- list_CLU_folders(inpath = "C:/Bep/Projects/CLUMondo/SSP Implementation/Scenario_", scenario = "SSP2")

neighmat <- "C:/Bep/Projects/CLUMondo/Model documentation/Model_codes_templates/neighmat.txt"

for(i in 1:length(CLU_folders)){
 folder <- CLU_folders[i]

 file.copy(alloc2, folder, overwrite = TRUE)
}
}
