% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/List_CLU_region_files.R
\name{list_CLU_region_files}
\alias{list_CLU_region_files}
\title{List the locations of CLU files}
\usage{
list_CLU_region_files(inpath, file)
}
\arguments{
\item{x}{the path to the folder containing the CLU folders}

\item{y}{the file name}
}
\value{
The locations of the files as a list
}
\description{
List the locations of CLU files
}
